sync - change to encoding/decoding and versioning - no longer prepending/slicing bytearrays for versions in addition to protobuf array usage - versions now stored as integers on models calling version routing components. -- SyncIntents and feature payloads no longer hold their version in their own bytearray, replaced with schema versions on serializable classes and the Intent model itself. Memory usage should be less crazy?

+++ core/contract/src/commonMain/kotlin/com/mochame/contract/exceptions/MochaExceptions.kt
@@ -61,7 +61,7 @@ sealed class MochaException(
         class Uncategorized(message: String? = null, cause: Throwable? = null) :
             Persistent(message ?: "Unexpected failure.", cause)
 
-        class UnknownProtocolVersion(version: Byte) :
+        class UnknownProtocolVersion(version: Int) :
             Persistent("Unknown protocol version: $version")
 
         class FileNotFound(blobId: String) :
